requirements/requirements.txt
setuptools
keyring
Django
psycopg
#psycopg2
psycopg2-binary
djangorestframework-simplejwt==5.2.2
djangorestframework
python-dotenv
drf_spectacular
redis
django-redis